feat: Move utils package to create a general one
