+++ pkg/s3-proxy/server/hostrouter.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/go-chi/chi/v5"
 	"github.com/gobwas/glob"
-	"github.com/oxyno-zeta/s3-proxy/pkg/s3-proxy/server/utils"
+	"github.com/oxyno-zeta/s3-proxy/pkg/s3-proxy/utils"
 )
 
 // Fork dead project https://github.com/go-chi/hostrouter/
