--- pkg/s3-proxy/server/utils/doc.go
@@ -1,3 +1,0 @@
-package utils
-
-// Main utils server
