+++ pkg/s3-proxy/utils/doc.go
@@ -1,0 +1,3 @@
+package utils
+
+// Main utils
