+++ pkg/s3-proxy/utils/mock_test.go
