+++ pkg/s3-proxy/utils/utils.go
