+++ pkg/s3-proxy/utils/utils_test.go
